Add SONiC Host Service package and infrastructure
This change adds infrastructure to create a host-side server that will
respond to specific requests over D-Bus. The intention is to let clients
that run within a container to query the host for the registered
requests. This also includes infrastructure in Translib to request the
host to perform some service.

This is primarily used for implementing actions that must be done in the
host system, such as image management, show techsupport, ZTP, etc.